requirements.txt
#django
Django==4.1.7
django-extensions==3.2.1
django-countries==7.5.1
# django-debug-toolbar==4.0.0
django-widget-tweaks==1.4.12
#python
python-decouple
pymemcache==4.0.0
python-memcached==1.59

#flake8
flake8==6.0.0

#postgresql
psycopg2-binary

#celery
celery==5.2.7
django-celery-beat==2.5.0

#payment
stripe==5.4.0

#other
django-cities-light
django-ajax-selects
django-phonenumber-field
django-phonenumbers
matplotlib
babel
pre-commit